+++ src/common/constants/constants.js
@@ -46,3 +46,5 @@ export const UNOPEN_ORDER_ABANDONED_HOLDINGS_TYPES = {
 
 export const FIELD_ARRAY_ITEM_IDENTIFIER_KEY = '__key__';
 export const OKAPI_TENANT_HEADER = 'X-Okapi-Tenant';
+
+export const HIDDEN_ORDER_FIELDS_FOR_VERSION_HISTORY = ['nextPolNumber'];