+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Optimize linked instances query to improve performance. Refs UIOR-1181.
 * Add permission to fetch invoices-related fiscal years into the orders "view" permission set. UIOR-1185.
 * Do not show special fields in "Version history" for PO and POL. Refs UIOR-1117.
+* Add validation for the `claimingInterval` field. Refs UIOR-1192.
 
 ## [5.0.1](https://github.com/folio-org/ui-orders/tree/v5.0.1) (2023-11-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v5.0.0...v5.0.1)