+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-orders
 
-## (5.1.0 IN PROGRESS)
+## (6.0.0 IN PROGRESS)
 
 * Allow Editing of "Renewal Date" and "Subscription to" on open purchase order line. Refs UIOR-1078.
 * Disable donor field in POL. Refs UIOR-1149.
@@ -12,6 +12,7 @@
 * Add permission to fetch invoices-related fiscal years into the orders "view" permission set. UIOR-1185.
 * Do not show special fields in "Version history" for PO and POL. Refs UIOR-1117.
 * Add validation for the `claimingInterval` field. Refs UIOR-1192.
+* *BREAKING* Add settings components for custom fields. Refs UIOR-1177.
 
 ## [5.0.1](https://github.com/folio-org/ui-orders/tree/v5.0.1) (2023-11-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v5.0.0...v5.0.1)